requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
albucore==0.0.23
albumentations==1.4.24
annotated-types==0.7.0
anyio==4.8.0
apache-beam==2.63.0
appdirs==1.4.4
array_record==0.6.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
autocommand==2.2.2
automated-interpretability==0.0.6
av==14.2.0
babe==0.0.7
backports.tarfile==1.2.0
beartype==0.14.1
beautifulsoup4==4.13.3
better-abc==0.0.3
bitsandbytes==0.45.1
blobfile==2.1.1
boostedblob==0.15.6
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cheroot==10.0.1
CherryPy==18.10.0
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==2.2.1
config2py==0.1.36
contourpy==1.3.1
cramjam==2.9.1
crcmod==1.7
cryptography==44.0.1
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.6.7
datasets==2.21.0
deep-translator==1.11.4
dill==0.3.1.1
distro==1.9.0
dm-tree==0.1.9
dnspython==2.7.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
docx==0.2.4
dol==0.3.9
eindex-callum==0.1.2
einops==0.7.0
etils==1.12.1
ExifRead-nocycle==3.0.1
fancy-einsum==0.0.3
fastavro==1.10.0
fasteners==0.19
fastparquet==2024.11.0
feedparser==6.0.11
filelock==3.17.0
fire==0.5.0
flash_attn==2.3.3
flatbuffers==25.2.10
fonttools==4.55.8
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.3.1
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
googletrans==4.0.2
gprof2dot==2024.6.6
graze==0.1.29
grpcio==1.65.5
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hdfs==2.7.3
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.28.1
hyperframe==6.1.0
i2==0.1.46
idna==3.10
imageio==2.37.0
img2dataset==1.45.0
immutabledict==4.2.1
importlib_metadata==8.0.0
importlib_resources==6.5.2
inflect==7.3.1
iniconfig==2.0.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jaxtyping==0.2.37
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==2.15.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
libretranslatepy==2.1.1
lxml==4.9.4
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
mkl-service==2.4.0
ml-dtypes==0.2.0
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
objsize==0.7.1
openai==1.62.0
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.15
pandas==2.2.3
patsy==1.0.1
pattern==0.0.1a0
pattern3==3.0.0
pdfminer.six==20240706
pdfminer3k==1.3.4
peft==0.15.0
pillow==11.1.0
plotly==5.24.1
plotly-express==0.4.1
pluggy==1.5.0
ply==3.11
portend==3.2.0
promise==2.3
propcache==0.2.1
proto-plus==1.26.0
protobuf==3.20.3
py2store==0.1.20
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocoevalcap==1.2
pycocotools==2.0.8
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydot==1.4.2
pymongo==4.11.2
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.3.4
pytest-profiling==1.8.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.0.0
qwen-vl-utils==0.0.10
rank-bm25==0.2.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9.1
sae-vis==0.3.6
safetensors==0.4.5
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
sgmllib3k==1.0.0
shellingham==1.5.4
shortuuid==1.0.13
simple-parsing==0.1.7
simplejson==3.19.3
simsimd==6.2.1
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
statsmodels==0.14.4
stringzilla==3.11.3
swig==4.3.0
sympy==1.13.1
tabulate==0.9.0
tempora==5.8.0
tenacity==9.0.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-datasets==4.9.7
tensorflow-estimator==2.15.0
tensorflow-hub==0.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-text==2.15.0
termcolor==2.5.0
tf_keras==2.19.0
threadpoolctl==3.5.0
tifffile==2025.1.10
tiktoken==0.6.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
torch==2.4.1
torchvision==0.16.1
tqdm==4.67.1
transformers==4.49.0
transformers-stream-generator==0.0.5
translate==3.6.1
triton==3.0.0
typeguard==4.4.1
typer==0.12.5
typing-inspect==0.9.0
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uvloop==0.21.0
wadler_lindig==0.1.3
wandb==0.16.6
webdataset==0.2.111
Werkzeug==3.1.3
Wikipedia-API==0.8.1
wrapt==1.14.1
xxhash==3.5.0
yarl==1.18.3
zc.lockfile==3.0.post1
zipp==3.21.0
zstandard==0.22.0

